tick event is working
