+++ src/Server/actions/board.js
@@ -6,12 +6,12 @@ import {
   noneToBoard,
   originToBoard,
 } from "../../utils/board";
-import { resetStats } from "../../utils/card";
-import { pushToReveal } from "../../utils/menu";
+import TickEvents from "../../models/tick/tick-events";
 import MenuRevealData from "../../models/menu/menu-reveal";
 import VersusContent from "../../models/versus-content";
+import { resetStats } from "../../utils/card";
+import { pushToReveal } from "../../utils/menu";
 import { safeSplice } from "../../utils/help";
-import Ticks from "../../models/tick/tick";
 
 const checkSelection = (G, selected, player) => {
   let origin = Object.keys(selected.origin)[0];
@@ -87,8 +87,7 @@ const tickCard = (G, ctx, place, index = 0) => {
   let card = getTileCard(G.board, place, index);
   card.status.forEach((tick, index) => {
     if (tick.duration > 0) {
-      //TODO TICK-EVENT
-      //Ticks[tick.name.toUpperCase()].event(card);
+      TickEvents[tick.name.toUpperCase()](card);
       tick.duration--;
       if (tick.duration === 0) {
         card.status = safeSplice(card.status, index);
