--- src/models/tick/burn.js
@@ -1,11 +1,0 @@
-import Tick from "./tick-object";
-import icon from "../../Client/ui/images/ticks/fire.png";
-import { TickType } from "./tick-object";
-
-const event = (card) => {
-  card.hp_mod--;
-};
-
-const Burn = Tick("Burn", TickType.AFFLICTIONS, icon, "Lose 1 HP Every TICK");
-
-export default Burn;
