--- src/models/tick/cooldown.js
@@ -1,14 +1,0 @@
-import Tick from "./tick-object";
-import icon from "../../Client/ui/images/ticks/cooldown.png";
-import { TickType } from "./tick-object";
-
-export const event = (card) => {};
-
-const Cooldown = Tick(
-  "Cooldown",
-  TickType.NEUTRAL,
-  icon,
-  "This card cannot apply it effect again until this reaches zero"
-);
-
-export default Cooldown;
