--- src/models/tick/stats-up.js
@@ -1,14 +1,0 @@
-import Tick from "./tick-object";
-import icon from "../../Client/ui/images/ticks/stats-up.png";
-import { TickType } from "./tick-object";
-
-const event = (card) => {};
-
-const StatusUp = Tick(
-  "Status Up",
-  TickType.BLESSINGS,
-  icon,
-  "This card status are temporary augmented by an effect."
-);
-
-export default StatusUp;
